SEQTaRget: Sequential Trial Emulation

Implementation of sequential trial emulation for the analysis of observational databases. The 'SEQTaRget' software accommodates time-varying treatments and confounders, as well as binary and failure time outcomes. 'SEQTaRget' allows to compare both static and dynamic strategies, can be used to estimate observational analogs of intention-to-treat and per-protocol effects, and can adjust for potential selection bias induced by losses-to-follow-up. (Paper to come).

See also

Author

Maintainer: Ryan O'Dea ryan.odea@psi.ch (ORCID)

Authors:

Ryan O'Dea ryan.odea@psi.ch (ORCID)

Alejandro Szmulewicz aszmulewicz@hsph.harvard.edu (ORCID)

Tom Palmer remlapmot@hotmail.com (ORCID) (ROR)

Paul Madley-Dowd p.madley-dowd@bristol.ac.uk (ORCID)

Miguel Hernán mhernan@hsph.harvard.edu (ORCID)

Other contributors:

The President and Fellows of Harvard College (ROR) [copyright holder]